requirements.txt
torch
torchvision
torchaudio
transformers
faiss-cpu
streamlit
ftfy
regex
tqdm
openai-whisper
scikit-learn
pandas
numpy
pytest
fastapi
uvicorn
python-multipart
opencv-python
git+https://github.com/openai/CLIP.git
